style(clippy): fix warnings in ff-core, ff-discovery, ff-terminal build_version

+++ crates/ff-terminal/build_version.rs
@@ -98,7 +98,7 @@ fn ymd_from_unix(secs: i64) -> (i32, u32, u32) {
     // Howard Hinnant's civil-date algorithm.
     let z = days + 719_468;
     let era = if z >= 0 { z } else { z - 146_096 } / 146_097;
-    let doe = (z - era * 146_097) as i64;
+    let doe = z - era * 146_097;
     let yoe = (doe - doe / 1460 + doe / 36_524 - doe / 146_096) / 365;
     let y = yoe + era * 400;
     let doy = doe - (365 * yoe + yoe / 4 - yoe / 100);
